anonymous viewing of posts and refresh working properly? (#26)
* We can reload pages without issue (NOT SHARE LINKS)

Before this commit, when we reloaded a page it would cause a white screen. Now that does not happen.

What im thinking is making it so that there's a "sandbox" page with just the comments and post.

* FIXED ANONOMOUS VIEWING

Changed by looking at the package and manifest paths. Make them absolute instead of relative.

---------

Co-authored-by: Brayan Quevedo Ramos <brayanquevedo2134@gmail.com>