WIP: Checking if email in use
Putting this here since I ran out of reads and writes from bad code (I was doing 100 reads per seconds whoops!).

Anyways, technically now the email does check to see whether the email is in use, but it always gives the good-to-go sign even on emails that are CLEARLY in use. I believe this is a firebase issue that I shall investigate later.