Testing test, all changes brayan and i have made so far (#30)
* We can reload pages without issue (NOT SHARE LINKS)

Before this commit, when we reloaded a page it would cause a white screen. Now that does not happen.

What im thinking is making it so that there's a "sandbox" page with just the comments and post.

* FIXED ANONOMOUS VIEWING

Changed by looking at the package and manifest paths. Make them absolute instead of relative.

* Added margin to the hexagon box on login

* Login.js now reloads for 2FA

So instead of extending the hexagon box, it'll just delete the login and password form and replace it with the 2FA form

* Finished Sign in and 2FA

Now we auto focus and send the verification code instead of having you do everything manually.

NOW onto working on Forgot Passowrd!

* Fixed reset Password field and more Login things!

* Inputs for Reset Password changed to match Login.js
* Now onEnterKeyPress works for 2FA and login
* Send Verification Code button has 10 second disabled if one was sent
* Changed 2FA placeholder names

* Force dark theme on Auth/Login/Sgnup/Forgot Password

* Revert "Force dark theme on Auth/Login/Sgnup/Forgot Password"

This reverts commit ff060266438983ec585b8ab85b2f6b619488c500.

* Began CSS changes to Signup

Need to finish Phone CSS and input validation

* Fixed styling on Signup.js

Now, we will break this down into three seperate parts.

Personal Info:
Firstname, lastname, email

Account Details:
Handle, Password, Confirm Password

Verification:
Phone Number, Verification code

* WIP: Finished Step1/2 of Signup.,js

It's a mess in here, working and doing my best

* New signup.js works!

Its vvvvv buggy, especially during the recaptcha portion, but will work on it later....

* Force page reload if navigate to login.js

This helps recaptcha reset and not cause any issues

* Back button now exists for third screen

Still need to change styling for everything since it looks kinda bad.

We can sucsessfully put in as many phone numbers as we wish, but if one is invalid since it DNE, then everything breaks??????????

* Added warning modal

So, if the user inputs a failing number, we're going to gaslight them by telling them it's a security issue.

* Verification Code fields disabled and enabled

Depends on the contex.

Disabled by default.
If the user gets a code, they enable.
Once the code have been verified, disable once again.

NOTE THIS BREAKS IF YOU HAVE TO DO A RECAPTCHA TEST

* Fixed Bug

Where if you navigated back from the last section, some things would break tehe

* FIXED PhoneInput CSS

* Fixed page 2 margin and padding!

* Changed ForgotPass to Hexagon

* Added Comments to ForgotPass.js

* Added comments for Signup.js

* FIX: Checking for Unique Handle

Now, the signup form checks to see whether the user handle is unique.

There might be a slight second between inputting an invalid handle and the next button being disabled but thats ok!

* WIP: Checking if email in use

Putting this here since I ran out of reads and writes from bad code (I was doing 100 reads per seconds whoops!).

Anyways, technically now the email does check to see whether the email is in use, but it always gives the good-to-go sign even on emails that are CLEARLY in use. I believe this is a firebase issue that I shall investigate later.

* Finished Email not in use

SO firebase automatically checks whether the user has an email in use, but that error is only thrown out when the user presses "sign-up".

So no changes need to happen here

* - added 2 files to use mentions but couldnt get it
  working
- Fixed orientation of post buttons, hovering, and color
- Fixed hovering of tags in posts
- Fixed theme in conversations and notifications
- Fixed lazy loading bug
- Fixed notifications so that it doesnt give you a notification if you
 comment or like your own post
- Fixed name for posts in notifications
- Tried fixing conversation bug to scroll to bottom of conversation
when u open it but could figure it out
- Removed 2 conversation related files that were unused

---------

Co-authored-by: Brayan Quevedo Ramos <brayanquevedo2134@gmail.com>